+++ test/e2e/storage/BUILD
@@ -104,6 +104,7 @@ filegroup(
         ":package-srcs",
         "//test/e2e/storage/drivers:all-srcs",
         "//test/e2e/storage/external:all-srcs",
+        "//test/e2e/storage/podlogs:all-srcs",
         "//test/e2e/storage/testpatterns:all-srcs",
         "//test/e2e/storage/testsuites:all-srcs",
         "//test/e2e/storage/utils:all-srcs",
