+++ test/e2e/storage/podlogs/BUILD
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["podlogs.go"],
-    importpath = "k8s.io/kubernetes/test/e2e/framework/podlogs",
+    importpath = "k8s.io/kubernetes/test/e2e/storage/podlogs",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
